+++ src/test/java/uk/co/jackoftradesltd/middle/player/PlayerBirthSellStatTest.java
@@ -262,10 +262,10 @@ class PlayerBirthSellStatTest {
      * @return the stored birth gold
      * @throws ReflectiveOperationException if the field cannot be reached
      */
-    private int auBirth() throws ReflectiveOperationException {
+    private long auBirth() throws ReflectiveOperationException {
         Field field = Player.class.getDeclaredField("auBirth");
         field.setAccessible(true);
-        return field.getInt(player);
+        return field.getLong(player);
     }
 
     /**
@@ -562,7 +562,7 @@ class PlayerBirthSellStatTest {
         @Test
         @DisplayName("false calls neither the birthpoints signal nor recalculateStats")
         void doesNeitherWhenClear() throws ReflectiveOperationException {
-            int before = auBirth();
+            long before = auBirth();
             Map<Stats, Integer> stats = statsMap(12, 13, 9, 12, 11);
             Map<Stats, Integer> spent = statsMap(2, 2, 0, 1, 3);
             Map<Stats, Integer> inc = statsMap(42, 42, 42, 42, 42);
